Expose VPL internals as well.
